Move users service implementation to internal package
