+++ api/internal/users/v1/internal/users/common_errors.go
@@ -1,4 +1,4 @@
-package v1
+package users
 
 import (
 	"fmt"
