+++ api/internal/users/v1/internal/users/connect_rpc_handler_test.go
@@ -1,4 +1,4 @@
-package v1
+package users
 
 import (
 	"buf.build/gen/go/fjarm/fjarm/connectrpc/go/fjarm/users/v1/usersv1connect"
