+++ api/internal/users/v1/internal/users/user_test.go
@@ -1,4 +1,4 @@
-package v1
+package users
 
 import (
 	"errors"
